[#49] refactor : Policy 인터페이스가 추후 여러 기능 패키지에서 사용될 것을 생각해, Common으로 위치 변경
