+++ src/main/java/com/aorri2/goodsforyou/common/policy/Policy.java
@@ -1,4 +1,4 @@
-package com.aorri2.goodsforyou.user.domain;
+package com.aorri2.goodsforyou.common.policy;
 
 public interface Policy<T> {
 	void apply(T t);
