+++ src/main/java/com/aorri2/goodsforyou/user/domain/NewUserPolicy.java
@@ -1,5 +1,6 @@
 package com.aorri2.goodsforyou.user.domain;
 
+import com.aorri2.goodsforyou.common.policy.Policy;
 import com.aorri2.goodsforyou.user.application.command.CreateUserCommand;
 
 public interface NewUserPolicy extends Policy<CreateUserCommand> {
